+++ packages/at_secondary_server/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.0.51
+- feat: Introduce option to unrevoke revoked enrollments
+- feat: Introduce option to delete enrollments that are denied/revoked
+- build[deps]: update dependency versions of at_commons -> 4.1.2, at_utils -> 3.0.18, at_lookup -> 3.0.48
 ## 3.0.50
 - fix: Enhance namespace authorisation check to verify when namespace has a period in it
 ## 3.0.49